requirements.txt
django==4.1.2
channels==4.0.0
whitenoise==6.2.0
daphne==4.0.0
uvloop==0.17.0
requests==2.26.0
psycopg2==2.9.5
django-allauth==0.52.0
lightstreamer-client==0.1
schedule==1.1.0
django_extensions==3.2.1